python-producer/pyproject.toml
[project]
name = "task-publisher"
version = "0.1.0"
description = "Python based task producer for rabbitMQ that generates and publishes data aggregation tasks to be processed by the Node.js backend service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pika>=1.3.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
